feat(frontend): improve homepage styles
