+++ src/pages/index.js
@@ -2,23 +2,25 @@ import React from "react";
 import Layout from "@theme/Layout";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 
-import { StartResourceSection } from "./startResource/index";
-import { HeroSection } from "./heroSection/index";
-import { MediumSection } from "./CustomArticle";
+import HeroSection from "./heroSection/index";
+import StartResourceSection from "./startResource";
+import MediumSection from "./articleSection";
 
 export const HomePage = () => {
   const { siteConfig } = useDocusaurusContext();
 
   return (
-    <Layout
-      permalink="/"
-      title={siteConfig.title}
-      description={siteConfig.tagline}
-    >
-      <HeroSection />
-      <StartResourceSection />
-      <MediumSection />
-    </Layout>
+    <>
+      <Layout
+        permalink="/"
+        title={siteConfig.title}
+        description={siteConfig.tagline}
+      >
+        <HeroSection />
+        <StartResourceSection />
+        <MediumSection />
+      </Layout>
+    </>
   );
 };
